Replace account specific passphrase with global passphrase
Storing the passphrase to decrypt a DB payload in the same table is less than optimal. The passphrase is now provided via a ENV variable and passed in via replicated. This is still not perfect, but already more secure than the current implementation. This single passphrase might be set as an in-memory variable some time in the future.