Connect account frontend to account setup (submission of keys to bank)
This adds a new field which stores the submission date of account data to bank servers. This way it should be easy to filter for which accounts we have to check for activation. Moreover, we have some more info about an account and it might be easier to detect if an account takes too long to get activated.