Refactor user and management authentication
In order to prepare for OAuth authentication, we need to refactor the way we authenticate users. First of all, we will have the concept of admin users in the near future. This won't work with our current concept of management tokens for organisations. Therefore we drop it and replace it with admin users.

In preparation of another authentication scheme, the authentication part is removed from our business logic and moved to a rack middleware. It provides user and organization data via the environment. This not only is much cleaner, but also allows us to scope uniqueness tests to organisations, as we only have access to the ENV in validations