Implement OAuth authentication middleware
This allows us to remove registration from the box and delegate it to a service which can handle it much better. Access code management will be handled there as well and the user can be presented with a proper UX for all those things. Setup of the box happens auto-magically as we can validate that the user's token is correct and all data verified.