Fix issues with invalid statements and refactor the way we import bank statements
We never really implemented the concept of an account statement. Those statements consist of multiple transactions in a given time and includes a sequence number, an opening and a closing balance. By embracing this kind of data and creating a dedicated data structure for it, we can simplify some of our internals. I.e. we can store the raw data of imports in those bank statements. Moreover, we can link each statement to those and rebuild them at any time.

In a future version, we could expose an endpoint to expose those bank statements, so other systems can import them and work with its data.