fix: add http status 500 as response
In swagger we did not define 500 as a valid response and they current
response was not returning `json` in the body.
This broke the swagger UI and could break other clients as well.

Note: To prevent any exception exploiting we only revels the error to
the configured error tracking system and the server logs.