refactor: use `sepa_file_parser` instead of `camt_parser`
we used a forked `camt_parser` version which was outdated.
To use the `sepa_file_parser` gem we extended the mapping factory in
order to keep the code impact as small as possible.

NOTE: thanks to the mapping the transaction_id is now better found which
caused a few spec to break due to invalid expectations. Those were
removed.